Move rules handling to its own module
There should otherwise be no code changes here.
